requirements.txt
db-dtypes==1.1.1
Flask==2.0.3
google-api-core==2.11.1
google-auth==2.21.0
google-cloud-bigquery==3.11.3
google-cloud-core==2.3.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pandas==2.0.3
requests==2.31.0
rsa==4.9
gunicorn